[refactor] 도메인 병합
Login -> Family 도메인으로 병합
기존 Login 도메인 코드 Family 도메인으로 이전
