+++ src/main/java/com/suitecare/suitecare/api/family/controller/FamilyController.java
@@ -1,7 +1,9 @@
-package com.suitecare.suitecare.api.controller;
+package com.suitecare.suitecare.api.family.controller;
 
-import com.suitecare.suitecare.api.service.FamilyService;
+import com.suitecare.suitecare.api.family.service.FamilyService;
 import com.suitecare.suitecare.domain.CreateFamilyDTO;
+import com.suitecare.suitecare.domain.LoginDTO;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,5 +22,16 @@ public class FamilyController {
     @GetMapping("/family")
     public int checkDuplicateID(@RequestParam String id){
         return familyService.checkDuplicateID(id);
+    }
+
+    @PostMapping("/login")
+    public void login(HttpServletResponse response, @RequestBody LoginDTO loginDTO) {
+        System.out.println(familyService.loginFamily(loginDTO));
+        if(familyService.loginFamily(loginDTO) != null) {
+            response.setStatus(200);
+            response.addHeader("msg", "success");
+        } else {
+            response.addHeader("msg", "fail");
+        }
     }
 }
