+++ src/main/java/com/suitecare/suitecare/api/family/service/FamilyService.java
@@ -1,7 +1,9 @@
-package com.suitecare.suitecare.api.service;
+package com.suitecare.suitecare.api.family.service;
 
-import com.suitecare.suitecare.api.repository.FamilyMapper;
+import com.suitecare.suitecare.api.family.repository.FamilyMapper;
 import com.suitecare.suitecare.domain.CreateFamilyDTO;
+import com.suitecare.suitecare.domain.FamilyDTO;
+import com.suitecare.suitecare.domain.LoginDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,5 +20,9 @@ public class FamilyService {
 
     public int checkDuplicateID(String id){
         return familyMapper.checkDuplicateID(id);
+    }
+
+    public FamilyDTO loginFamily(LoginDTO login){
+        return familyMapper.login(login);
     }
 }
