+++ src/main/resources/mapper/FamilyMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.suitecare.suitecare.api.repository.FamilyMapper">
+<mapper namespace="com.suitecare.suitecare.api.family.repository.FamilyMapper">
 
     <!-- 회원가입 -->
     <insert id="createFamily" parameterType="CreateFamilyDTO">
